+++ tests/echoserver.yaml.sh
@@ -113,6 +113,7 @@ spec:
       annotations:
         sidecar-injector.cyberark.com/inject: "yes"
         sidecar-injector.cyberark.com/secretlessConfig: "k8s/crd#crd-basic-auth-proxy"
+        sidecar-injector.cyberark.com/secretlessCRDSuffix: ${SECRETLESS_CRD_SUFFIX}
         sidecar-injector.cyberark.com/injectType: "secretless"
 
     spec: